server/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
sqlalchemy==2.0.28
psycopg[binary]==3.1.18
pyjwt==2.8.0
python-dotenv==1.0.1
